requirements.txt
torch
torchvision
torch-directml
streamlit
opencv-contrib-python
Pillow
numpy
matplotlib
seaborn
scikit-learn
